PYIC-7982: refactor to use eventResolver
