+++ lambdas/process-journey-event/src/main/java/uk/gov/di/ipv/core/processjourneyevent/exceptions/JourneyEngineException.java
@@ -5,6 +5,10 @@ public class JourneyEngineException extends Exception {
         super();
     }
 
+    public JourneyEngineException(String message) {
+        super(message);
+    }
+
     public JourneyEngineException(String message, Throwable cause) {
         super(message, cause);
     }
