+++ lambdas/process-journey-event/src/main/java/uk/gov/di/ipv/core/processjourneyevent/statemachine/events/BasicEvent.java
@@ -1,36 +1,17 @@
 package uk.gov.di.ipv.core.processjourneyevent.statemachine.events;
 
 import lombok.Data;
-import org.apache.logging.log4j.LogManager;
-import org.apache.logging.log4j.Logger;
-import software.amazon.awssdk.utils.StringUtils;
 import uk.gov.di.ipv.core.library.auditing.AuditEventTypes;
 import uk.gov.di.ipv.core.library.domain.IpvJourneyTypes;
-import uk.gov.di.ipv.core.library.domain.ScopeConstants;
-import uk.gov.di.ipv.core.library.exceptions.CiExtractionException;
-import uk.gov.di.ipv.core.library.exceptions.ConfigException;
-import uk.gov.di.ipv.core.library.exceptions.CredentialParseException;
-import uk.gov.di.ipv.core.library.helpers.LogHelper;
-import uk.gov.di.ipv.core.library.persistence.item.ClientOAuthSessionItem;
-import uk.gov.di.ipv.core.processjourneyevent.exceptions.JourneyEngineException;
-import uk.gov.di.ipv.core.processjourneyevent.statemachine.TransitionResult;
-import uk.gov.di.ipv.core.processjourneyevent.statemachine.exceptions.MissingSecurityCheckCredential;
-import uk.gov.di.ipv.core.processjourneyevent.statemachine.exceptions.UnknownEventException;
 import uk.gov.di.ipv.core.processjourneyevent.statemachine.states.JourneyChangeState;
 import uk.gov.di.ipv.core.processjourneyevent.statemachine.states.State;
 
-import java.text.ParseException;
 import java.util.LinkedHashMap;
 import java.util.List;
 import java.util.Map;
-import java.util.Optional;
-
-import static uk.gov.di.ipv.core.library.config.ConfigurationVariable.CREDENTIAL_ISSUER_ENABLED;
-import static uk.gov.di.ipv.core.library.domain.ErrorResponse.MISSING_SECURITY_CHECK_CREDENTIAL;
 
 @Data
 public class BasicEvent implements Event {
-    private static final Logger LOGGER = LogManager.getLogger();
     private String name;
     private String targetJourney;
     private String targetState;
@@ -42,111 +23,6 @@ public class BasicEvent implements Event {
     private List<AuditEventTypes> auditEvents;
     private LinkedHashMap<String, String> auditContext;
     private LinkedHashMap<String, Event> checkMitigation;
-
-    public TransitionResult resolve(EventResolveParameters resolveParameters)
-            throws UnknownEventException, JourneyEngineException {
-        try {
-            var journeyContext = resolveParameters.journeyContext();
-            var configService = resolveParameters.configService();
-
-            if (checkIfDisabled != null) {
-                Optional<String> firstDisabledCri =
-                        checkIfDisabled.keySet().stream()
-                                .filter(
-                                        id ->
-                                                !configService.getBooleanParameter(
-                                                        CREDENTIAL_ISSUER_ENABLED, id))
-                                .findFirst();
-                if (firstDisabledCri.isPresent()) {
-                    String disabledCriId = firstDisabledCri.get();
-                    LOGGER.info(
-                            "CRI with ID '{}' is disabled. Using alternative event", disabledCriId);
-                    return checkIfDisabled.get(disabledCriId).resolve(resolveParameters);
-                }
-            }
-            if (checkJourneyContext != null && !StringUtils.isEmpty(journeyContext)) {
-                Optional<String> matchingContext =
-                        checkJourneyContext.keySet().stream()
-                                .filter(ctx -> ctx.equals(journeyContext))
-                                .findFirst();
-                if (matchingContext.isPresent()) {
-                    String contextValue = matchingContext.get();
-                    LOGGER.info(
-                            "Matching context '{}' is set. Using alternative event", contextValue);
-                    return checkJourneyContext.get(contextValue).resolve(resolveParameters);
-                }
-            }
-            if (checkFeatureFlag != null) {
-                Optional<String> firstFeatureFlag =
-                        checkFeatureFlag.keySet().stream()
-                                .filter(configService::enabled)
-                                .findFirst();
-                if (firstFeatureFlag.isPresent()) {
-                    String featureFlagValue = firstFeatureFlag.get();
-                    LOGGER.info(
-                            "Feature flag '{}' is set. Using alternative event", featureFlagValue);
-                    return checkFeatureFlag.get(featureFlagValue).resolve(resolveParameters);
-                }
-            }
-
-            if (isCheckMitigationAllowed(resolveParameters.clientOAuthSessionItem())) {
-                var matchedMitigation = getMitigationEvent(resolveParameters);
-
-                if (matchedMitigation.isPresent()) {
-                    var mitigationEvent = checkMitigation.get(matchedMitigation.get());
-                    LOGGER.info(
-                            "Mitigation '{}' found. Using alternative event.",
-                            matchedMitigation.get());
-                    return mitigationEvent.resolve(resolveParameters);
-                }
-            }
-
-            return new TransitionResult(
-                    targetStateObj, auditEvents, auditContext, targetEntryEvent);
-        } catch (MissingSecurityCheckCredential
-                | CiExtractionException
-                | CredentialParseException
-                | ParseException
-                | ConfigException e) {
-            throw new JourneyEngineException("Failed to resolve event", e);
-        }
-    }
-
-    private boolean isCheckMitigationAllowed(ClientOAuthSessionItem clientOAuthSessionItem) {
-        var isReverification =
-                !StringUtils.isEmpty(clientOAuthSessionItem.getScope())
-                        && clientOAuthSessionItem
-                                .getScopeClaims()
-                                .contains(ScopeConstants.REVERIFICATION);
-        return checkMitigation != null && !isReverification;
-    }
-
-    private Optional<String> getMitigationEvent(EventResolveParameters resolveParameters)
-            throws MissingSecurityCheckCredential, CiExtractionException, CredentialParseException,
-                    ParseException, ConfigException {
-        var ipvSessionItem = resolveParameters.ipvSessionItem();
-        var clientOAuthSessionItem = resolveParameters.clientOAuthSessionItem();
-        var cimitUtilityService = resolveParameters.cimitUtilityService();
-
-        var securityCheckCredential = ipvSessionItem.getSecurityCheckCredential();
-
-        if (StringUtils.isEmpty(securityCheckCredential)) {
-            LOGGER.error(LogHelper.buildErrorMessage(MISSING_SECURITY_CHECK_CREDENTIAL));
-            throw new MissingSecurityCheckCredential("Missing security check credential");
-        }
-
-        var contraIndicators =
-                cimitUtilityService.getContraIndicatorsFromVc(
-                        securityCheckCredential, clientOAuthSessionItem.getUserId());
-
-        var validMitigation =
-                cimitUtilityService.getMitigationJourneyEvent(
-                        contraIndicators, ipvSessionItem.getTargetVot());
-
-        return (validMitigation.isPresent() && checkMitigation.containsKey(validMitigation.get()))
-                ? validMitigation
-                : Optional.empty();
-    }
 
     @Override
     public void initialize(
