+++ lambdas/process-journey-event/src/main/java/uk/gov/di/ipv/core/processjourneyevent/statemachine/events/Event.java
@@ -2,9 +2,6 @@ package uk.gov.di.ipv.core.processjourneyevent.statemachine.events;
 
 import com.fasterxml.jackson.annotation.JsonSubTypes;
 import com.fasterxml.jackson.annotation.JsonTypeInfo;
-import uk.gov.di.ipv.core.processjourneyevent.exceptions.JourneyEngineException;
-import uk.gov.di.ipv.core.processjourneyevent.statemachine.TransitionResult;
-import uk.gov.di.ipv.core.processjourneyevent.statemachine.exceptions.UnknownEventException;
 import uk.gov.di.ipv.core.processjourneyevent.statemachine.states.State;
 
 import java.util.Map;
@@ -16,9 +13,6 @@ import java.util.Map;
     @JsonSubTypes.Type(value = ExitNestedJourneyEvent.class)
 })
 public interface Event {
-
-    TransitionResult resolve(EventResolveParameters resolveParameters)
-            throws UnknownEventException, JourneyEngineException;
 
     void initialize(
             String name, Map<String, State> states, Map<String, Event> nestedJourneyExitEvents);
