+++ lambdas/process-journey-event/src/main/java/uk/gov/di/ipv/core/processjourneyevent/statemachine/states/BasicState.java
@@ -8,6 +8,7 @@ import uk.gov.di.ipv.core.processjourneyevent.exceptions.JourneyEngineException;
 import uk.gov.di.ipv.core.processjourneyevent.statemachine.TransitionResult;
 import uk.gov.di.ipv.core.processjourneyevent.statemachine.events.Event;
 import uk.gov.di.ipv.core.processjourneyevent.statemachine.events.EventResolveParameters;
+import uk.gov.di.ipv.core.processjourneyevent.statemachine.events.EventResolver;
 import uk.gov.di.ipv.core.processjourneyevent.statemachine.exceptions.UnknownEventException;
 import uk.gov.di.ipv.core.processjourneyevent.statemachine.exceptions.UnknownStateException;
 import uk.gov.di.ipv.core.processjourneyevent.statemachine.stepresponses.StepResponse;
@@ -32,21 +33,25 @@ public class BasicState implements State {
 
     @Override
     public TransitionResult transition(
-            String eventName, String startState, EventResolveParameters eventResolveParameters)
+            String eventName,
+            String startState,
+            EventResolveParameters eventResolveParameters,
+            EventResolver eventResolver)
             throws UnknownEventException, UnknownStateException, JourneyEngineException {
         // Special recovery event
         if (ATTEMPT_RECOVERY_EVENT.equals(eventName)) {
             return new TransitionResult(this);
         }
 
-        return getEvent(eventName)
-                .orElseThrow(
-                        () ->
-                                new UnknownEventException(
-                                        String.format(
-                                                "Unknown event provided to '%s' state: '%s'",
-                                                name, eventName)))
-                .resolve(eventResolveParameters);
+        var event =
+                getEvent(eventName)
+                        .orElseThrow(
+                                () ->
+                                        new UnknownEventException(
+                                                String.format(
+                                                        "Unknown event provided to '%s' state: '%s'",
+                                                        name, eventName)));
+        return eventResolver.resolve(event, eventResolveParameters);
     }
 
     private Optional<Event> getEvent(String eventName) {
