+++ lambdas/process-journey-event/src/main/java/uk/gov/di/ipv/core/processjourneyevent/statemachine/states/JourneyChangeState.java
@@ -5,6 +5,7 @@ import lombok.Data;
 import uk.gov.di.ipv.core.library.domain.IpvJourneyTypes;
 import uk.gov.di.ipv.core.processjourneyevent.statemachine.TransitionResult;
 import uk.gov.di.ipv.core.processjourneyevent.statemachine.events.EventResolveParameters;
+import uk.gov.di.ipv.core.processjourneyevent.statemachine.events.EventResolver;
 import uk.gov.di.ipv.core.processjourneyevent.statemachine.exceptions.UnknownEventException;
 import uk.gov.di.ipv.core.processjourneyevent.statemachine.exceptions.UnknownStateException;
 
@@ -16,7 +17,10 @@ public class JourneyChangeState implements State {
 
     @Override
     public TransitionResult transition(
-            String eventName, String startState, EventResolveParameters eventResolveParameters)
+            String eventName,
+            String startState,
+            EventResolveParameters eventResolveParameters,
+            EventResolver eventResolver)
             throws UnknownEventException, UnknownStateException {
         throw new IllegalStateException("Cannot transition from JourneyChangeState");
     }
