+++ lambdas/process-journey-event/src/main/java/uk/gov/di/ipv/core/processjourneyevent/statemachine/states/NestedJourneyInvokeState.java
@@ -10,6 +10,7 @@ import uk.gov.di.ipv.core.processjourneyevent.exceptions.JourneyEngineException;
 import uk.gov.di.ipv.core.processjourneyevent.statemachine.TransitionResult;
 import uk.gov.di.ipv.core.processjourneyevent.statemachine.events.Event;
 import uk.gov.di.ipv.core.processjourneyevent.statemachine.events.EventResolveParameters;
+import uk.gov.di.ipv.core.processjourneyevent.statemachine.events.EventResolver;
 import uk.gov.di.ipv.core.processjourneyevent.statemachine.exceptions.UnknownEventException;
 import uk.gov.di.ipv.core.processjourneyevent.statemachine.exceptions.UnknownStateException;
 
@@ -34,7 +35,10 @@ public class NestedJourneyInvokeState implements State {
 
     @Override
     public TransitionResult transition(
-            String eventName, String startState, EventResolveParameters eventResolveParameters)
+            String eventName,
+            String startState,
+            EventResolveParameters eventResolveParameters,
+            EventResolver eventResolver)
             throws UnknownEventException, UnknownStateException, JourneyEngineException {
         Queue<String> stateNameParts = getStateNameParts(startState);
 
@@ -47,7 +51,7 @@ public class NestedJourneyInvokeState implements State {
                                 "Unknown entry event '%s' for '%s' state nested journey definition",
                                 eventName, name));
             }
-            result = event.resolve(eventResolveParameters);
+            result = eventResolver.resolve(event, eventResolveParameters);
         } else {
             stateNameParts.remove();
             State currentNestedState =
@@ -62,7 +66,8 @@ public class NestedJourneyInvokeState implements State {
                     currentNestedState.transition(
                             eventName,
                             String.join(JOURNEY_STATE_DELIMITER, stateNameParts),
-                            eventResolveParameters);
+                            eventResolveParameters,
+                            eventResolver);
         }
 
         if (result.state() instanceof NestedJourneyInvokeState) {
@@ -71,7 +76,8 @@ public class NestedJourneyInvokeState implements State {
                     .transition(
                             entryEvent,
                             String.join(JOURNEY_STATE_DELIMITER, stateNameParts),
-                            eventResolveParameters);
+                            eventResolveParameters,
+                            eventResolver);
         }
 
         return result;
