+++ lambdas/process-journey-event/src/test/java/uk/gov/di/ipv/core/processjourneyevent/statemachine/StateMachineTest.java
@@ -1,13 +1,13 @@
 package uk.gov.di.ipv.core.processjourneyevent.statemachine;
 
 import org.junit.jupiter.api.Test;
+import org.mockito.Mock;
 import uk.gov.di.ipv.core.library.auditing.AuditEventTypes;
 import uk.gov.di.ipv.core.library.domain.JourneyState;
 import uk.gov.di.ipv.core.library.persistence.item.ClientOAuthSessionItem;
 import uk.gov.di.ipv.core.library.persistence.item.IpvSessionItem;
-import uk.gov.di.ipv.core.library.service.CimitUtilityService;
-import uk.gov.di.ipv.core.library.service.ConfigService;
 import uk.gov.di.ipv.core.processjourneyevent.statemachine.events.EventResolveParameters;
+import uk.gov.di.ipv.core.processjourneyevent.statemachine.events.EventResolver;
 import uk.gov.di.ipv.core.processjourneyevent.statemachine.exceptions.UnknownStateException;
 import uk.gov.di.ipv.core.processjourneyevent.statemachine.states.BasicState;
 import uk.gov.di.ipv.core.processjourneyevent.statemachine.states.NestedJourneyDefinition;
@@ -30,18 +30,16 @@ import static uk.gov.di.ipv.core.library.domain.IpvJourneyTypes.INITIAL_JOURNEY_
 class StateMachineTest {
     private static final EventResolveParameters EVENT_RESOLVE_PARAMETERS =
             new EventResolveParameters(
-                    "journeyContext",
-                    mock(ConfigService.class),
-                    new IpvSessionItem(),
-                    new ClientOAuthSessionItem(),
-                    new CimitUtilityService(mock(ConfigService.class)));
+                    "journeyContext", new IpvSessionItem(), new ClientOAuthSessionItem());
+    @Mock private EventResolver eventResolver;
 
     @Test
     void transitionShouldReturnAppropriateState() throws Exception {
         var expectedResult = new TransitionResult(new BasicState());
 
         State startingState = mock(BasicState.class);
-        when(startingState.transition("event", "START_STATE", EVENT_RESOLVE_PARAMETERS))
+        when(startingState.transition(
+                        "event", "START_STATE", EVENT_RESOLVE_PARAMETERS, eventResolver))
                 .thenReturn(expectedResult);
 
         StateMachineInitializer mockStateMachineInitializer = mock(StateMachineInitializer.class);
@@ -51,7 +49,8 @@ class StateMachineTest {
         StateMachine stateMachine = new StateMachine(mockStateMachineInitializer);
 
         var actualResult =
-                stateMachine.transition("START_STATE", "event", null, EVENT_RESOLVE_PARAMETERS);
+                stateMachine.transition(
+                        "START_STATE", "event", null, EVENT_RESOLVE_PARAMETERS, eventResolver);
 
         assertEquals(expectedResult, actualResult);
     }
@@ -68,18 +67,24 @@ class StateMachineTest {
                 UnknownStateException.class,
                 () ->
                         stateMachine.transition(
-                                "UNKNOWN_STATE", "event", null, EVENT_RESOLVE_PARAMETERS));
+                                "UNKNOWN_STATE",
+                                "event",
+                                null,
+                                EVENT_RESOLVE_PARAMETERS,
+                                eventResolver));
     }
 
     @Test
     void transitionShouldTransitionIntoNestedJourneyInvokeState() throws Exception {
         var expectedResult = new TransitionResult(new BasicState());
         State nestedJourneyInvokeState = mock(NestedJourneyInvokeState.class);
-        when(nestedJourneyInvokeState.transition("event", "START_STATE", EVENT_RESOLVE_PARAMETERS))
+        when(nestedJourneyInvokeState.transition(
+                        "event", "START_STATE", EVENT_RESOLVE_PARAMETERS, eventResolver))
                 .thenReturn(expectedResult);
 
         State startingState = mock(BasicState.class);
-        when(startingState.transition("event", "START_STATE", EVENT_RESOLVE_PARAMETERS))
+        when(startingState.transition(
+                        "event", "START_STATE", EVENT_RESOLVE_PARAMETERS, eventResolver))
                 .thenReturn(new TransitionResult(nestedJourneyInvokeState));
 
         StateMachineInitializer mockStateMachineInitializer = mock(StateMachineInitializer.class);
@@ -89,7 +94,8 @@ class StateMachineTest {
         StateMachine stateMachine = new StateMachine(mockStateMachineInitializer);
 
         var actualResult =
-                stateMachine.transition("START_STATE", "event", null, EVENT_RESOLVE_PARAMETERS);
+                stateMachine.transition(
+                        "START_STATE", "event", null, EVENT_RESOLVE_PARAMETERS, eventResolver);
 
         assertEquals(expectedResult, actualResult);
     }
@@ -100,11 +106,13 @@ class StateMachineTest {
         // Arrange
         var nestedResult = new TransitionResult(new BasicState());
         State nestedJourneyInvokeState = mock(NestedJourneyInvokeState.class);
-        when(nestedJourneyInvokeState.transition("event", "START_STATE", EVENT_RESOLVE_PARAMETERS))
+        when(nestedJourneyInvokeState.transition(
+                        "event", "START_STATE", EVENT_RESOLVE_PARAMETERS, eventResolver))
                 .thenReturn(nestedResult);
 
         State startingState = mock(BasicState.class);
-        when(startingState.transition("event", "START_STATE", EVENT_RESOLVE_PARAMETERS))
+        when(startingState.transition(
+                        "event", "START_STATE", EVENT_RESOLVE_PARAMETERS, eventResolver))
                 .thenReturn(
                         new TransitionResult(
                                 nestedJourneyInvokeState,
@@ -120,7 +128,8 @@ class StateMachineTest {
 
         // Act
         var actualResult =
-                stateMachine.transition("START_STATE", "event", null, EVENT_RESOLVE_PARAMETERS);
+                stateMachine.transition(
+                        "START_STATE", "event", null, EVENT_RESOLVE_PARAMETERS, eventResolver);
 
         // Assert
         var expectedResult =
@@ -143,11 +152,13 @@ class StateMachineTest {
                         Map.of("testKey", "testValue"),
                         null);
         State nestedJourneyInvokeState = mock(NestedJourneyInvokeState.class);
-        when(nestedJourneyInvokeState.transition("event", "START_STATE", EVENT_RESOLVE_PARAMETERS))
+        when(nestedJourneyInvokeState.transition(
+                        "event", "START_STATE", EVENT_RESOLVE_PARAMETERS, eventResolver))
                 .thenReturn(nestedResult);
 
         State startingState = mock(BasicState.class);
-        when(startingState.transition("event", "START_STATE", EVENT_RESOLVE_PARAMETERS))
+        when(startingState.transition(
+                        "event", "START_STATE", EVENT_RESOLVE_PARAMETERS, eventResolver))
                 .thenReturn(new TransitionResult(nestedJourneyInvokeState));
 
         StateMachineInitializer mockStateMachineInitializer = mock(StateMachineInitializer.class);
@@ -158,7 +169,8 @@ class StateMachineTest {
 
         // Act
         var actualResult =
-                stateMachine.transition("START_STATE", "event", null, EVENT_RESOLVE_PARAMETERS);
+                stateMachine.transition(
+                        "START_STATE", "event", null, EVENT_RESOLVE_PARAMETERS, eventResolver);
 
         // Assert
         var expectedResult =
@@ -181,11 +193,13 @@ class StateMachineTest {
                         Map.of("testKey1", "testValue1"),
                         null);
         State nestedJourneyInvokeState = mock(NestedJourneyInvokeState.class);
-        when(nestedJourneyInvokeState.transition("event", "START_STATE", EVENT_RESOLVE_PARAMETERS))
+        when(nestedJourneyInvokeState.transition(
+                        "event", "START_STATE", EVENT_RESOLVE_PARAMETERS, eventResolver))
                 .thenReturn(nestedResult);
 
         State startingState = mock(BasicState.class);
-        when(startingState.transition("event", "START_STATE", EVENT_RESOLVE_PARAMETERS))
+        when(startingState.transition(
+                        "event", "START_STATE", EVENT_RESOLVE_PARAMETERS, eventResolver))
                 .thenReturn(
                         new TransitionResult(
                                 nestedJourneyInvokeState,
@@ -201,7 +215,8 @@ class StateMachineTest {
 
         // Act
         var actualResult =
-                stateMachine.transition("START_STATE", "event", null, EVENT_RESOLVE_PARAMETERS);
+                stateMachine.transition(
+                        "START_STATE", "event", null, EVENT_RESOLVE_PARAMETERS, eventResolver);
 
         // Assert
         var expectedResult =
@@ -225,12 +240,13 @@ class StateMachineTest {
         var expectedResult = new TransitionResult(new BasicState());
         State nestedJourneyInvokeState = mock(NestedJourneyInvokeState.class);
         when(nestedJourneyInvokeState.transition(
-                        eventOverride, "START_STATE", EVENT_RESOLVE_PARAMETERS))
+                        eventOverride, "START_STATE", EVENT_RESOLVE_PARAMETERS, eventResolver))
                 .thenReturn(expectedResult);
 
         // Outer event transitions using the overall event
         State startingState = mock(BasicState.class);
-        when(startingState.transition(event, "START_STATE", EVENT_RESOLVE_PARAMETERS))
+        when(startingState.transition(
+                        event, "START_STATE", EVENT_RESOLVE_PARAMETERS, eventResolver))
                 .thenReturn(
                         new TransitionResult(nestedJourneyInvokeState, null, null, eventOverride));
 
@@ -241,7 +257,8 @@ class StateMachineTest {
         StateMachine stateMachine = new StateMachine(mockStateMachineInitializer);
 
         var actualResult =
-                stateMachine.transition("START_STATE", event, null, EVENT_RESOLVE_PARAMETERS);
+                stateMachine.transition(
+                        "START_STATE", event, null, EVENT_RESOLVE_PARAMETERS, eventResolver);
 
         assertEquals(expectedResult, actualResult);
     }
@@ -252,7 +269,10 @@ class StateMachineTest {
 
         State startingState = mock(NestedJourneyInvokeState.class);
         when(startingState.transition(
-                        "event", "START_STATE/NESTED_JOURNEY", EVENT_RESOLVE_PARAMETERS))
+                        "event",
+                        "START_STATE/NESTED_JOURNEY",
+                        EVENT_RESOLVE_PARAMETERS,
+                        eventResolver))
                 .thenReturn(expectedResult);
 
         StateMachineInitializer mockStateMachineInitializer = mock(StateMachineInitializer.class);
@@ -263,7 +283,11 @@ class StateMachineTest {
 
         var actualResult =
                 stateMachine.transition(
-                        "START_STATE/NESTED_JOURNEY", "event", null, EVENT_RESOLVE_PARAMETERS);
+                        "START_STATE/NESTED_JOURNEY",
+                        "event",
+                        null,
+                        EVENT_RESOLVE_PARAMETERS,
+                        eventResolver);
 
         assertEquals(expectedResult, actualResult);
     }
