+++ lambdas/process-journey-event/src/test/java/uk/gov/di/ipv/core/processjourneyevent/statemachine/states/NestedJourneyInvokeStateTest.java
@@ -1,13 +1,15 @@
 package uk.gov.di.ipv.core.processjourneyevent.statemachine.states;
 
 import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.ExtendWith;
+import org.mockito.Mock;
+import org.mockito.junit.jupiter.MockitoExtension;
 import uk.gov.di.ipv.core.library.persistence.item.ClientOAuthSessionItem;
 import uk.gov.di.ipv.core.library.persistence.item.IpvSessionItem;
-import uk.gov.di.ipv.core.library.service.CimitUtilityService;
-import uk.gov.di.ipv.core.library.service.ConfigService;
 import uk.gov.di.ipv.core.processjourneyevent.statemachine.TransitionResult;
 import uk.gov.di.ipv.core.processjourneyevent.statemachine.events.BasicEvent;
 import uk.gov.di.ipv.core.processjourneyevent.statemachine.events.EventResolveParameters;
+import uk.gov.di.ipv.core.processjourneyevent.statemachine.events.EventResolver;
 import uk.gov.di.ipv.core.processjourneyevent.statemachine.exceptions.UnknownEventException;
 import uk.gov.di.ipv.core.processjourneyevent.statemachine.exceptions.UnknownStateException;
 
@@ -20,20 +22,19 @@ import static org.mockito.ArgumentMatchers.eq;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
 
+@ExtendWith(MockitoExtension.class)
 class NestedJourneyInvokeStateTest {
     private static final EventResolveParameters EVENT_RESOLVE_PARAMETERS =
             new EventResolveParameters(
-                    "journeyContext",
-                    mock(ConfigService.class),
-                    new IpvSessionItem(),
-                    new ClientOAuthSessionItem(),
-                    new CimitUtilityService(mock(ConfigService.class)));
+                    "journeyContext", new IpvSessionItem(), new ClientOAuthSessionItem());
+    @Mock private EventResolver eventResolver;
 
     @Test
     void transitionShouldUseEntryEventsWhenStartStateHasOnePart() throws Exception {
         var expectedResult = new TransitionResult(new BasicState());
         BasicEvent basicEvent = mock(BasicEvent.class);
-        when(basicEvent.resolve(any(EventResolveParameters.class))).thenReturn(expectedResult);
+        when(eventResolver.resolve(eq(basicEvent), any(EventResolveParameters.class)))
+                .thenReturn(expectedResult);
 
         NestedJourneyDefinition nestedJourneyDefinition = new NestedJourneyDefinition();
         nestedJourneyDefinition.setEntryEvents(Map.of("next", basicEvent));
@@ -43,7 +44,7 @@ class NestedJourneyInvokeStateTest {
 
         var actualResult =
                 nestedJourneyInvokeState.transition(
-                        "next", "INVOKE_STATE", EVENT_RESOLVE_PARAMETERS);
+                        "next", "INVOKE_STATE", EVENT_RESOLVE_PARAMETERS, eventResolver);
 
         assertEquals(expectedResult, actualResult);
     }
@@ -57,7 +58,10 @@ class NestedJourneyInvokeStateTest {
 
         var expectedResult = new TransitionResult(new BasicState());
         when(currentNestedState.transition(
-                        eq("next"), eq("NESTED_STATE"), any(EventResolveParameters.class)))
+                        eq("next"),
+                        eq("NESTED_STATE"),
+                        any(EventResolveParameters.class),
+                        eq(eventResolver)))
                 .thenReturn(expectedResult);
 
         NestedJourneyInvokeState nestedJourneyInvokeState = new NestedJourneyInvokeState();
@@ -65,7 +69,10 @@ class NestedJourneyInvokeStateTest {
 
         var actualResult =
                 nestedJourneyInvokeState.transition(
-                        "next", "INVOKE_STATE/NESTED_STATE", EVENT_RESOLVE_PARAMETERS);
+                        "next",
+                        "INVOKE_STATE/NESTED_STATE",
+                        EVENT_RESOLVE_PARAMETERS,
+                        eventResolver);
 
         assertEquals(expectedResult, actualResult);
     }
@@ -83,15 +90,24 @@ class NestedJourneyInvokeStateTest {
                 mock(NestedJourneyInvokeState.class);
         var expectedResult = new TransitionResult(new BasicState());
         when(nestedNestedJourneyInvokeState.transition(
-                        eq("next"), eq("NESTED_STATE"), any(EventResolveParameters.class)))
+                        eq("next"),
+                        eq("NESTED_STATE"),
+                        any(EventResolveParameters.class),
+                        eq(eventResolver)))
                 .thenReturn(expectedResult);
         when(currentNestedState.transition(
-                        eq("next"), eq("NESTED_STATE"), any(EventResolveParameters.class)))
+                        eq("next"),
+                        eq("NESTED_STATE"),
+                        any(EventResolveParameters.class),
+                        eq(eventResolver)))
                 .thenReturn(new TransitionResult(nestedNestedJourneyInvokeState));
 
         var actualResult =
                 nestedJourneyInvokeState.transition(
-                        "next", "INVOKE_STATE/NESTED_STATE", EVENT_RESOLVE_PARAMETERS);
+                        "next",
+                        "INVOKE_STATE/NESTED_STATE",
+                        EVENT_RESOLVE_PARAMETERS,
+                        eventResolver);
 
         assertEquals(expectedResult, actualResult);
     }
@@ -108,7 +124,10 @@ class NestedJourneyInvokeStateTest {
                 UnknownEventException.class,
                 () ->
                         nestedJourneyInvokeState.transition(
-                                "unknown", "INVOKE_STATE", EVENT_RESOLVE_PARAMETERS));
+                                "unknown",
+                                "INVOKE_STATE",
+                                EVENT_RESOLVE_PARAMETERS,
+                                eventResolver));
     }
 
     @Test
@@ -124,7 +143,10 @@ class NestedJourneyInvokeStateTest {
                 UnknownStateException.class,
                 () ->
                         nestedJourneyInvokeState.transition(
-                                "unknown", "INVOKE_STATE/UNKNOWN_STATE", EVENT_RESOLVE_PARAMETERS));
+                                "unknown",
+                                "INVOKE_STATE/UNKNOWN_STATE",
+                                EVENT_RESOLVE_PARAMETERS,
+                                eventResolver));
     }
 
     @Test
